add dialog-esc-composable (chromatic-build)
