+++ packages/vlossom/src/components/vs-layout/VsLayout.scss
@@ -1,6 +1,7 @@
 .vs-layout {
     position: relative;
     width: 100%;
+    height: 100%;
     overflow: hidden;
     transition:
         padding-left 0.3s,
